[prim_sha2,rtl] Add RTL implementation + update core + lint waivers
This adds the RTL implementation for the merged SHA-2 prim that can be
configured at compile-time as either multi-mode or 256-bit.
The RTL includes the 32-bit word input wrapper prim_sha2_32 that
instantiates the prim_sha2. This commit also updates the relevant prim
core files as well as the lint waivers.

Signed-off-by: Ghada Dessouky <gdessouky@google.com>